Ultimo commint

+++ src/domain/models/Order.kt
@@ -1,11 +1,13 @@
 package domain.models
 
+import java.util.Date
+
 data class Order(
     val idOrder: String,
     val nombreCliente: String,
-    val dni: Byte,
+    val dni: Int,
     val total: Double,
-    val fecha: String,
+    val fecha:String,
     val tipoOperacion: String,
     val codigoSede: String
 )
